extracts main tab into separate component
