+++ src/components/GeneralDashboard/NavigationForm.jsx
@@ -15,21 +15,9 @@ import {
 import useCommonStyles from '../../theme/useCommonStyles'
 import AppData from '../../modules/AppData'
 import SubmitButton from '../SubmitButton'
-import { withStyles } from '@material-ui/core/styles'
+import MainTabFormSection from './MainTabFormSection'
 
-const StyledSwitch = withStyles({
-  switchBase: {
-    color: '#ddd',
-    '&$checked': {
-      color: '#0BDA51',
-    },
-    '&$checked + $track': {
-      backgroundColor: '#00FF00',
-    },
-  },
-  checked: {},
-  track: {},
-})(Switch)
+
 
 const NavigationForm = ({ mainTabs, secondaryTabs }) => {
   const [expanded, setExpanded] = useState({ footer: true })
@@ -45,7 +33,7 @@ const NavigationForm = ({ mainTabs, secondaryTabs }) => {
   }
 
   const mainTabsList = mainTabs.map((mainTab, index) => {
-    let secondaryTabList = (<></>)
+    let secondaryTabList = <></>
     if (mainTab.secondary_tabs) {
       secondaryTabList = mainTab.secondary_tabs.map((secondaryTab) => (
         <h1>{secondaryTab.label}</h1>
@@ -53,51 +41,9 @@ const NavigationForm = ({ mainTabs, secondaryTabs }) => {
     }
 
     return (
-      <Grid item container direction='row' alignItems='center'>
-        <Typography variant='h7' style={{ width: '2rem' }}>
-          {index}.
-        </Typography>
-        <Controller
-          name={`main-tab-${index}-input`}
-          control={control}
-          defaultValue={mainTab.label}
-          rules={{ required: 'This field cannot be empty' }}
-          render={({ field: { onChange, value }, fieldState: { error } }) => (
-            <TextField
-              data-cy={`main-tab-${index}-input`}
-              variant='outlined'
-              label={`Label*`}
-              inputProps={{ maxLength: labelMaxLength }}
-              error={!!error}
-              helperText={error ? error.message : null}
-              value={value}
-              onChange={onChange}
-            />
-          )}
-        />
-        <Controller
-          name={`main-tab-${1}-switch`}
-          control={control}
-          defaultValue={mainTab.visible}
-          render={({ field: { onChange, value } }) => (
-            <FormControlLabel
-              control={
-                <StyledSwitch
-                  checked={value}
-                  onChange={onChange}
-                  articleId={1}
-                />
-              }
-              label={
-                <Typography className={commonClasses.switchLabel}>
-                  {value ? 'Visible' : 'Hidden'}
-                </Typography>
-              }
-              labelPlacement='bottom'
-            />
-          )}
-        />
-        {secondaryTabList && secondaryTabList}
+      <Grid item container direction='column'>
+        <MainTabFormSection index={index} label={mainTab.label} visible={mainTab.visible}/>
+        <Grid item>{secondaryTabList && secondaryTabList}</Grid>
       </Grid>
     )
   })
